+++ yash-syntax/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `yash-syntax` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ????-??-??
+
+### Changed
+
+- `syntax::HereDoc::content` redefined as `RefCell<Text>` (previously `Text`)
+
 ## [0.2.0] - 2022-02-03
 
 Previously, source code attribution attached to ASTs was line-oriented. The